+++ app/cmd/rpc/client.go
@@ -5,14 +5,12 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"net/http"
-	"strings"
-	"time"
-
 	"github.com/julienschmidt/httprouter"
 	"github.com/pokt-network/pocket-core/app"
 	"github.com/pokt-network/pocket-core/x/pocketcore/types"
+	"io/ioutil"
+	"net/http"
+	"strings"
 )
 
 // Dispatch supports CORS functionality
@@ -43,14 +41,9 @@ type RPCRelayResponse struct {
 	Response  string `json:"response"`
 	// remove proof object because client already knows about it
 }
-func timeTrack(start time.Time, name string) {
-	elapsed := time.Since(start)
-	fmt.Printf("%s took %s\n", name, elapsed)
-}
 
 // Relay supports CORS functionality
 func Relay(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
-	defer timeTrack(time.Now(), "entire relay")
 	var relay = types.Relay{}
 	if cors(&w, r) {
 		return