+++ x/pocketcore/types/key.go
@@ -2,7 +2,6 @@ package types
 
 import (
 	sdk "github.com/pokt-network/posmint/types"
-	"time"
 )
 
 const (
@@ -82,7 +81,6 @@ func KeyForClaims(addr sdk.Address) ([]byte, error) {
 
 // "KeyForEvidence" - Generates the key for evidence
 func KeyForEvidence(header SessionHeader, evidenceType EvidenceType) ([]byte, error) {
-	defer timeTrack(time.Now(), "key for evidence")
 	// validate the evidence type
 	if evidenceType != RelayEvidence && evidenceType != ChallengeEvidence {
 		return nil, NewInvalidEvidenceErr(ModuleName)