+++ doc/changelog.md
@@ -46,6 +46,7 @@
 - Split DAOTx transfer & burn to their own CLI commands to avoid user prone errors
 - Added minimum proofs to pocketcore module params
 - Split burn and transfer DAO CLI command, avoid error prone errors
+- Cache flushes to the database periodically instead of per relay for efficiency
 
 ## RC-0.3.0
 - Added governance module from posmint